+++ st2common/st2common/exceptions/auth.py
