+++ st2common/st2common/hooks.py
@@ -23,7 +23,7 @@ from six.moves.urllib import parse as urlparse
 from webob import exc
 
 from st2common import log as logging
-from st2common.exceptions import access as exceptions
+from st2common.exceptions import auth as exceptions
 from st2common.util.jsonify import json_encode
 from st2common.util.auth import validate_token
 from st2common.constants.auth import HEADER_ATTRIBUTE_NAME
