+++ st2common/st2common/models/api/access.py
@@ -16,7 +16,7 @@
 from oslo.config import cfg
 from st2common.util import isotime
 from st2common.models.api.base import BaseAPI
-from st2common.models.db.access import UserDB, TokenDB
+from st2common.models.db.auth import UserDB, TokenDB
 from st2common import log as logging
